Se comienza a generar un home/index.html para las vistas

+++ cgi-cpn/prontus_cfg_save.cgi
@@ -634,22 +634,25 @@ sub validarPort {
             &glib_html_02::print_json_result(0, "No se pudo localizar la Plantilla de Portada que se utilizará en edición base [$prontus_varglb::PRONTUS_ID/plantillas/edic/nroedic/port/$item]", 'exit=1,ctype=1');
         };
     } elsif ($var eq 'PORT_HOME') {
-
-        #~ print STDERR "item[$item]\n";
-
         if (! -f $dir_portada_site . $item) {
             &glib_html_02::print_json_result(0, "No se pudo localizar la portada que se utilizará como página de inicio [$prontus_varglb::PRONTUS_ID/site/edic/base/port/$item]", 'exit=1,ctype=1');
         } else {
             # Crear link simbólico hacia la portada.
             my $path_home_index = $prontus_varglb::DIR_SERVER . '/' . $prontus_varglb::PRONTUS_ID . '/site/edic/base/home/index.html';
             my $cmd = "ln -s $dir_portada_site$item $path_home_index";
-
-            if (-f $path_home_index && -l $path_home_index) {
-                unlink $path_home_index;
-            };
-
-            #~ print STDERR "cmd[$cmd]\n";
+            unlink $path_home_index if (-f $path_home_index && -l $path_home_index);
             system $cmd;
+
+            my $mv_dir_portada_site = $dir_portada_site;
+            chop $mv_dir_portada_site;
+            foreach my $mv (keys %prontus_varglb::MULTIVISTAS) {
+                # Crear link simbólico hacia la portada por cada multivista
+                my $path_home_index = $prontus_varglb::DIR_SERVER . '/' . $prontus_varglb::PRONTUS_ID . '/site/edic/base/home/index_'.$mv.'.html';
+                my $path_home_port = "$mv_dir_portada_site-$mv/$item";
+                my $cmd = "ln -s $path_home_port $path_home_index";
+                unlink $path_home_index if (-f $path_home_index && -l $path_home_index);
+                system $cmd;
+            };
         };
     };
